Added uploading images/files and downloading pdf

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const {validationResult} = require ('express-validator');
 
+const fileHelper = require ('../util/file');
+
 const Product = require ('../models/product');
 
 exports.getAddProduct = (req, res) => {
@@ -13,22 +15,37 @@ exports.getAddProduct = (req, res) => {
   });
 };
 
-exports.postAddProduct = (req, res) => {
+exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if (!image) {
+    return res.status (422).render ('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description,
+      },
+      errorMessage: 'Attached file is not an image. Please only use .png, .jpg, .jpeg files',
+      validationErrors: [],
+    });
+  }
+
   const errors = validationResult (req);
 
   if (!errors.isEmpty ()) {
     return res.status (422).render ('admin/edit-product', {
       pageTitle: 'Add Product',
-      path: '/admin/edit-product',
+      path: '/admin/add-product',
       editing: false,
       hasError: true,
       product: {
         title,
-        imageUrl,
         price,
         description,
       },
@@ -36,6 +53,8 @@ exports.postAddProduct = (req, res) => {
       validationErrors: errors.array (),
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product ({
     title,
@@ -50,11 +69,13 @@ exports.postAddProduct = (req, res) => {
       res.redirect ('/admin/products');
     })
     .catch (err => {
-      console.log (err);
+      const error = new Error (err);
+      error.httpStatusCode = 500;
+      return next (error);
     });
 };
 
-exports.getEditProduct = (req, res) => {
+exports.getEditProduct = (req, res, next) => {
   const editMode = req.query.edit;
   if (!editMode) {
     return res.redirect ('/');
@@ -75,14 +96,18 @@ exports.getEditProduct = (req, res) => {
         validationErrors: [],
       });
     })
-    .catch (err => console.log (err));
+    .catch (err => {
+      const error = new Error (err);
+      error.httpStatusCode = 500;
+      return next (error);
+    });
 };
 
 exports.postEditProduct = (req, res) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDescription = req.body.description;
 
   const errors = validationResult (req);
@@ -95,7 +120,6 @@ exports.postEditProduct = (req, res) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
         _id: prodId,
@@ -113,22 +137,39 @@ exports.postEditProduct = (req, res) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile (product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save ().then (() => {
         res.redirect ('/admin/products');
       });
     })
-    .catch (err => console.log (err));
+    .catch (err => {
+      const error = new Error (err);
+      error.httpStatusCode = 500;
+      return next (error);
+    });
 };
 
-exports.postDeleteProduct = (req, res) => {
+exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne ({_id: prodId, userId: req.user._id})
+  Product.findById (prodId)
+    .then (product => {
+      if (!product) {
+        next (new Error ('Product not found'));
+      }
+      fileHelper.deleteFile (product.imageUrl);
+      return Product.deleteOne ({_id: prodId, userId: req.user._id});
+    })
     .then (() => {
-      console.log ('Deleted the product');
       res.redirect ('/admin/products');
     })
-    .catch (err => console.log (err));
+    .catch (err => {
+      const error = new Error (err);
+      error.httpStatusCode = 500;
+      return next (error);
+    });
 };
 
 exports.getProducts = (req, res) => {
@@ -140,5 +181,9 @@ exports.getProducts = (req, res) => {
         path: '/admin/products',
       });
     })
-    .catch (err => console.log (err));
+    .catch (err => {
+      const error = new Error (err);
+      error.httpStatusCode = 500;
+      return next (error);
+    });
 };
